requirements.txt
amqp==2.5.1
aniso8601==7.0.0
appnope==0.1.0
astroid==2.3.2
backcall==0.1.0
billiard==3.6.1.0
celery==4.3.0
coverage==4.5.4
decorator==4.4.1
Django==2.2.6
django-celery-beat==1.5.0
django-filter==2.2.0
django-timezone-field==3.1
djangorestframework==3.10.3
elasticsearch==6.4.0
elasticsearch-dsl==6.1.0
graphene==2.1.8
graphene-django==2.6.0
graphql-core==2.2.1
graphql-relay==2.0.0
gunicorn==19.9.0
importlib-metadata==0.23
ipaddress==1.0.23
ipdb==0.12.2
ipython==7.9.0
ipython-genutils==0.2.0
iso8601==0.1.12
isort==4.3.21
jedi==0.15.1
kombu==4.6.5
lazy-object-proxy==1.4.2
Markdown==3.1.1
mccabe==0.6.1
more-itertools==7.2.0
numpy==1.17.3
pandas==0.25.2
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
promise==2.2.1
prompt-toolkit==2.0.10
ptyprocess==0.6.0
Pygments==2.4.2
pylint==2.4.3
python-crontab==2.4.0
python-dateutil==2.8.0
python-json-logger==0.1.11
python-logstash==0.4.6
python3-logstash==0.4.80
pytz==2019.3
Rx==1.6.1
singledispatch==3.4.0.3
six==1.12.0
sqlparse==0.3.0
traitlets==4.3.3
typed-ast==1.4.0
typing==3.7.4.1
urllib3==1.25.6
vine==1.3.0
wcwidth==0.1.7
wrapt==1.11.2
zipp==0.6.0
